# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ecommerce API",
    "version": "1.0.0",
    "description": "API for searching products, viewing product details, and creating orders."
  },
  "servers": [
    {
      "url": "https://ecommerce-mcp.onrender.com"
    }
  ],
  "paths": {
    "/api/search-products": {
      "post": {
        "operationId": "searchProducts",
        "summary": "Search products",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "query": {
                    "type": "string"
                  }
                },
                "required": [
                  "query"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "List of products"
          }
        }
      }
    },
    "/api/products/{productId}": {
      "get": {
        "operationId": "getProduct",
        "summary": "Get product details",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product details"
          },
          "404": {
            "description": "Product not found"
          }
        }
      }
    },
    "/api/create-order": {
      "post": {
        "operationId": "createOrder",
        "summary": "Create an order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "integer"
                  },
                  "quantity": {
                    "type": "integer"
                  }
                },
                "required": [
                  "productId",
                  "quantity"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Order confirmation"
          }
        }
      }
    }
  }
}